+++ src/watchdog/events.py
@@ -88,8 +88,8 @@ Event Handler Classes
 import os.path
 import logging
 import re
-from pathtools.patterns import match_any_paths
 from watchdog.utils import has_attribute
+from watchdog.utils.patterns import match_any_paths
 
 
 EVENT_TYPE_MOVED = 'moved'