+++ changelog.rst
@@ -11,6 +11,7 @@ Changelog
 - Allow file paths on Unix that don't follow the file system encoding (`# <https://github.com/gorakhargosh/watchdog/pull/703>`_)
 - Drop support for Python 2.7 (`# <https://github.com/gorakhargosh/watchdog/pull/703>`_)
 - Thanks to our beloved contributors: @SamSchott
+- Use `pathlib` from the standard library, instead of pathtools: @bstaletic
 
 
 0.10.4